+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fixed
 
+- negative col start offset for raw put to statement which lives on compound block header line
 - use Modifying context manager on `get_slice(..., cut=True)`
 - recognize multiline parenthesized `ImportFrom.names` as enclosed
 - get one from field which can contain a value but is currently `None`